refactor export fallback e2e fixtures
