+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components-flat/app/layout.js
@@ -1,0 +1,7 @@
+export default function RootLayout({ children }) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  )
+}
