+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components-flat/next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  output: 'export',
+  trailingSlash: false,
+  cacheComponents: true,
+}
+
+module.exports = nextConfig
