+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/(grouped)/grouped/[slug]/slug-client.js
@@ -1,0 +1,9 @@
+'use client'
+
+import { useParams } from 'next/navigation'
+
+export default function GroupedSlugClient() {
+  const params = useParams()
+
+  return <h1>{params.slug}</h1>
+}
