+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/docs/[...slug]/slug-client.js
@@ -1,0 +1,11 @@
+'use client'
+
+import { useParams } from 'next/navigation'
+
+export default function DocsSlugClient() {
+  const params = useParams()
+
+  return (
+    <h1>{Array.isArray(params.slug) ? params.slug.join('/') : 'missing'}</h1>
+  )
+}
