+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/docs/page.js
@@ -1,0 +1,16 @@
+import Link from 'next/link'
+
+export default function DocsIndex() {
+  return (
+    <main>
+      <h1>Docs</h1>
+      <ul>
+        <li>
+          <Link href="/docs/guides/export/fallback">
+            docs export fallback page
+          </Link>
+        </li>
+      </ul>
+    </main>
+  )
+}
