+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/inbox/@modal/[thread]/modal-client.js
@@ -1,0 +1,9 @@
+'use client'
+
+import { useParams } from 'next/navigation'
+
+export default function InboxModalClient() {
+  const params = useParams()
+
+  return <p id="modal-thread">Modal {params.thread}</p>
+}
