+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/inbox/@modal/[thread]/page.js
@@ -1,0 +1,10 @@
+import { Suspense } from 'react'
+import InboxModalClient from './modal-client'
+
+export default function InboxModalPage() {
+  return (
+    <Suspense fallback={<p id="modal-thread">Loading modal...</p>}>
+      <InboxModalClient />
+    </Suspense>
+  )
+}
