+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/inbox/@modal/default.js
@@ -1,0 +1,3 @@
+export default function InboxModalDefault() {
+  return <p id="modal-thread">No modal</p>
+}
