+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/inbox/[thread]/thread-client.js
@@ -1,0 +1,9 @@
+'use client'
+
+import { useParams } from 'next/navigation'
+
+export default function InboxThreadClient() {
+  const params = useParams()
+
+  return <h1>{params.thread}</h1>
+}
