+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/inbox/layout.js
@@ -1,0 +1,8 @@
+export default function InboxLayout({ children, modal }) {
+  return (
+    <main>
+      {children}
+      <section id="modal-slot">{modal}</section>
+    </main>
+  )
+}
