+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/optional/[[...slug]]/slug-client.js
@@ -1,0 +1,10 @@
+'use client'
+
+import { useParams } from 'next/navigation'
+
+export default function OptionalSlugClient() {
+  const params = useParams()
+  const slug = params.slug
+
+  return <h1>{Array.isArray(slug) ? slug.join('/') : 'optional index'}</h1>
+}
