+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-cache-components/app/page.js
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <main>
+      <h1>Home</h1>
+    </main>
+  )
+}
