+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-known-params-cache-components/app/not-found.js
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  return <h1>My custom not found page</h1>
+}
