+++ test/e2e/app-dir-export/fixtures/dynamic-fallback-known-params-cache-components/next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  output: 'export',
+  trailingSlash: true,
+  cacheComponents: true,
+}
+
+module.exports = nextConfig
